Added MVV-LVA heuristic Results of 0.2.6 vs 0.2.5 (10+0.1, NULL, NULL, 8moves_v3.pgn): Elo: 53.53 +/- 17.87, nElo: 82.76 +/- 27.17 LOS: 100.00 %, DrawRatio: 40.76 %, PairsRatio: 2.32 Games: 628, Wins: 179, Losses: 83, Draws: 366, Points: 362.0 (57.64 %) Ptnml(0-2): [6, 50, 128, 102, 28], WL/DD Ratio: 0.20

+++ UCI Test/Move.cs
@@ -9,6 +9,6 @@
         public char PromPiece { get; set; }
         public byte LastCapture { get; set; }
         public int EnPassentIndex { get; set; }
-        public byte MoveValue { get; set; }
+        public int MoveValue { get; set; }
     }
 }
